+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/synthesis/evaluation/latency/SimpleLatencyEvaluation.java
@@ -34,8 +34,15 @@
  */
 package org.preesm.algorithm.synthesis.evaluation.latency;
 
+import java.util.Map;
 import org.preesm.algorithm.synthesis.SynthesisResult;
 import org.preesm.algorithm.synthesis.evaluation.ISynthesisEvaluator;
+import org.preesm.algorithm.synthesis.timer.ActorExecutionTiming;
+import org.preesm.algorithm.synthesis.timer.SimpleTimer;
+import org.preesm.model.pisdf.AbstractActor;
+import org.preesm.model.pisdf.PiGraph;
+import org.preesm.model.scenario.Scenario;
+import org.preesm.model.slam.Design;
 
 /**
  *
@@ -45,9 +52,16 @@ import org.preesm.algorithm.synthesis.evaluation.ISynthesisEvaluator;
 public class SimpleLatencyEvaluation implements ISynthesisEvaluator<LatencyCost> {
 
   @Override
-  public LatencyCost evaluate(final SynthesisResult synthesisChoice) {
-    // TODO
-    return new LatencyCost(0);
+  public LatencyCost evaluate(final PiGraph algo, final Design slamDesign, final Scenario scenario,
+      final SynthesisResult synthesisChoice) {
+
+    final Map<AbstractActor, ActorExecutionTiming> computeTimings = new SimpleTimer(algo,
+        synthesisChoice.schedule, synthesisChoice.mapping, scenario).computeTimings();
+
+    final long latency = computeTimings.entrySet().stream().mapToLong(entry -> entry.getValue().getEndTime()).max()
+        .orElse(0L);
+
+    return new LatencyCost(latency);
   }
 
 }